remove use of eigen large headers globally
