+++ src/aliceVision/lightingEstimation/lightingCalibration.cpp
@@ -25,6 +25,7 @@
 // Eigen
 #include <Eigen/Dense>
 #include <Eigen/Core>
+#include <Eigen/QR>
 
 #include <filesystem>
 #include <math.h>
