+++ src/aliceVision/multiview/relativePose/Essential8PSolver.cpp
@@ -9,6 +9,7 @@
 #include "Essential8PSolver.hpp"
 #include <aliceVision/multiview/epipolarEquation.hpp>
 #include <aliceVision/numeric/algebra.hpp>
+#include <Eigen/SVD>
 
 namespace aliceVision {
 namespace multiview {
