+++ src/aliceVision/multiview/relativePose/Fundamental7PSolver.cpp
@@ -10,6 +10,9 @@
 #include <aliceVision/numeric/polynomial.hpp>
 #include <aliceVision/multiview/epipolarEquation.hpp>
 #include <aliceVision/numeric/algebra.hpp>
+
+#include <Eigen/SVD>
+
 namespace aliceVision {
 namespace multiview {
 namespace relativePose {
