+++ src/aliceVision/numeric/NumericFunctions.cpp
@@ -1,12 +1,13 @@
 // This file is part of the AliceVision project.
-// Copyright (c) 2016 AliceVision contributors.
+// Copyright (c) 2025 AliceVision contributors.
 // Copyright (c) 2012 openMVG contributors.
 // Copyright (c) 2007 libmv contributors.
 // This Source Code Form is subject to the terms of the Mozilla Public License,
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "numeric.hpp"
+#include <aliceVision/numeric/numeric.hpp>
+#include "NumericFunctions.hpp"
 
 #include <iostream>
 #include <fstream>
