+++ src/aliceVision/sfm/pipeline/panorama/ReconstructionEngine_panorama.cpp
@@ -37,6 +37,7 @@
 #include <aliceVision/matching/supportEstimation.hpp>
 
 #include <dependencies/htmlDoc/htmlDoc.hpp>
+#include <Eigen/SVD>
 
 #ifdef _MSC_VER
     #pragma warning(once : 4267)  // warning C4267: 'argument' : conversion from 'size_t' to 'const int', possible loss of data
