+++ src/software/pipeline/main_checkerboardCalibration.cpp
@@ -36,6 +36,8 @@
 
 #include <boost/program_options.hpp>
 
+#include <Eigen/SVD>
+
 #include <fstream>
 #include <map>
 #include <utility>
